Make the best effort to create hash-based collections of requested size.
For instance, if a Set<Integer> of size 10 is requested, with integer range [1..10],
there is a high chance of collisions which may result in a set of size less than 10.
This commit address this edge case. It also handles the case where requested size
is impossible, e.g. Set<Boolean> of size 5.