Bug fix: propagate Map/Collection generator errors conditionally (#900)
`4.1.0` introduced a change where the generator propagates an error
if it fails to instantiate a map or collection (instead of returning null,
which was the original behaviour). This causes a regression for some users
This commit modifies the behaviour to propagates the error only if
`Keys.FAIL_ON_ERROR` is enabled.